Add utility functions and tests
This adds some utility functions for common functionality when
working with config files. It also adds some super basic tests
that tests those functionalities.

These tests at present rely on the host to have the Steam snap
installed with at least 1 game installed, and will modify the
host's Steam configs. In the future this needs to be mocked
instead.